refactor: リポジトリインターフェースをrepository/interfaces.goに集約 (#2075)
- NoteLinkRepositoryInterface を service/ から repository/interfaces.go に移動
- NoteTemplateRepositoryInterface を service/ から repository/interfaces.go に移動
- サービス層の参照を repository.XxxInterface に統一
- アーキテクチャ境界の一貫性を確保（全インターフェースがrepository層に集約）
- 全テスト通過確認済み
